feat: improve footer
